move xmark to separate file
