+++ lib/src/components/card_rounded/card_rounded.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 import 'package:help_ukraine_widget/src/components/card_rounded/close_button_card_rounded.dart';
+import 'package:help_ukraine_widget/src/components/xmark.dart';
 
 /// This is a card widget with a rounded border, that has optional close button
 /// at the top right.
@@ -107,63 +108,5 @@ class CardRounded extends StatelessWidget {
       clipBehavior: Clip.none,
       children: [card, _positionedCloseButton],
     );
-  }
-}
-
-/// XMark Icon
-class XMark extends StatelessWidget {
-	/// size of the canvas to draw a cross on.
-  final Size size;
-
-	/// color of the lines.
-  final Color color;
-  
-	/// width of cross lines.
-  final double lineWidth;
-
-  /// Constructor
-  const XMark({
-    Key? key,
-    required this.size,
-    required this.color,
-    required this.lineWidth,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return CustomPaint(	
-			size: size,
-			painter: _CrossPainter(lineWidth: lineWidth),
-    );
-  }
-}
-
-class _CrossPainter extends CustomPainter {
-  final double lineWidth;
-
-	/// constructor
-  _CrossPainter({ required this.lineWidth });
-
-  @override
-  void paint(Canvas canvas, Size size) { 
-    final paint = Paint()
-      ..color = Colors.white
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = lineWidth;
-    canvas.drawLine(
-			Offset.zero,
-			Offset(size.width, size.height),
-			paint,
-		);
-    canvas.drawLine(
-			Offset(size.width, 0),
-			Offset(0, size.height),
-			paint,
-		); 
-	}
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    return true;
   }
 }
